refactoring: extracting constants and base entity to separate files
